+++ e2e/tests/theme-default/sidebar.spec.ts
@@ -2,9 +2,9 @@ import { expect, test } from '@playwright/test'
 
 test.describe('has heading sidebar', () => {
   test('frontmatter', async ({ page }) => {
-    await page.goto('sidebar/auto.html')
+    await page.goto('sidebar/frontmatter/heading.html')
 
-    const sidebarItems = page.locator('a.sidebar-item')
+    const sidebarItems = page.locator('a.vp-sidebar-item')
 
     await expect(sidebarItems.nth(0)).toContainText('Sidebar Heading 1')
     await expect(sidebarItems.nth(1)).toContainText('Sidebar Heading 2')
@@ -13,9 +13,9 @@ test.describe('has heading sidebar', () => {
   test('config', async ({ page }) => {
     await page.goto('sidebar/heading/1.html')
 
-    const sidebarHeadings = page.locator('.sidebar-heading')
-    const pageTitle = page.locator('#vp-content h1')
-    const sidebarItems = page.locator('.sidebar-item.route-link')
+    const sidebarHeadings = page.locator('.vp-sidebar-heading')
+    const pageTitle = page.locator('.theme-default-content h1')
+    const sidebarItems = page.locator('.vp-sidebar-item.route-link')
 
     const pageTitleText = (await pageTitle.innerText()).replace('#', '')
 
@@ -30,11 +30,30 @@ test.describe('has heading sidebar', () => {
   })
 })
 
-test('has configured sidebar', async ({ page }) => {
-  await page.goto('sidebar/config/1.html')
+test.describe('has configured sidebar', () => {
+  test('theme config', async ({ page }) => {
+    await page.goto('sidebar/config/1.html')
 
-  const sidebarItems = page.locator('a.sidebar-item')
+    const sidebarItems = page.locator('a.vp-sidebar-item')
 
-  await expect(sidebarItems.nth(1)).toContainText('sidebar 1')
-  await expect(sidebarItems.nth(4)).toContainText('sidebar 2')
+    await expect(sidebarItems.nth(1)).toContainText('sidebar 1')
+    await expect(sidebarItems.nth(4)).toContainText('sidebar 2')
+  })
+
+  test('frontmatter', async ({ page }) => {
+    await page.goto('sidebar/frontmatter/config.html')
+
+    const sidebarItems = page.locator('a.vp-sidebar-item')
+
+    await expect(sidebarItems.nth(0)).toContainText('Home')
+    await expect(sidebarItems.nth(1)).toContainText('主页')
+  })
+})
+
+test('frontmatter disable', async ({ page }) => {
+  await page.goto('sidebar/frontmatter/disable.html')
+
+  const sidebarItems = page.locator('a.vp-sidebar-item')
+
+  expect(await sidebarItems.count()).toBe(0)
 })