+++ themes/theme-default/src/client/components/HomeContent.vue
@@ -1,5 +1,0 @@
-<template>
-  <div id="vp-content">
-    <Content />
-  </div>
-</template>