+++ themes/theme-default/src/node/defaultTheme.ts
@@ -62,6 +62,16 @@ export const defaultTheme = ({
             path.resolve(__dirname, '../client/components', file),
           ]),
       ),
+      // use alias to make all composables replaceable
+      ...Object.fromEntries(
+        fs
+          .readdirSync(path.resolve(__dirname, '../client/composables'))
+          .filter((file) => file.endsWith('.js'))
+          .map((file) => [
+            `@theme/${file.substring(0, file.length - 3)}`,
+            path.resolve(__dirname, '../client/composables', file),
+          ]),
+      ),
     },
 
     clientConfigFile: path.resolve(__dirname, '../client/config.js'),
@@ -77,7 +87,7 @@ export const defaultTheme = ({
       // @vuepress/plugin-active-header-link
       themePlugins.activeHeaderLinks !== false
         ? activeHeaderLinksPlugin({
-            headerLinkSelector: 'a.sidebar-item',
+            headerLinkSelector: 'a.vp-sidebar-item',
             headerAnchorSelector: '.header-anchor',
             // should greater than page transition duration
             delay: 300,