+++ plugins/blog/plugin-comment/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-rc.34](https://github.com/vuepress/ecosystem/compare/v2.0.0-rc.33...v2.0.0-rc.34) (2024-06-04)
+
+### Features
+
+- **plugin-comment:** export use apis ([6ac961f](https://github.com/vuepress/ecosystem/commit/6ac961f33f8211b01938323588faa49757d74516))
+
 # [2.0.0-rc.33](https://github.com/vuepress/ecosystem/compare/v2.0.0-rc.32...v2.0.0-rc.33) (2024-05-30)
 
 **Note:** Version bump only for package @vuepress/plugin-comment